requirements.txt
snowflake-snowpark-python[pandas]
pandas
plotly
matplotlib
seaborn
streamlit==1.32.0
snowflake-connector-python==3.10.0
pandas==2.2.2
plotly==5.19.0
numpy==1.26.4
python-dotenv==1.0.1